ai-service/src/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
requests==2.31.0
python-dotenv==1.0.0
